+++ packages/xchain-thornode/CHANGELOG.md
@@ -1,3 +1,9 @@
+# v0.3.4 (2023-07-26)
+
+## Update
+
+- Update thornode to the latest THORNode 1.116.0
+
 # v0.3.3 (2023-06-21)
 
 ## Update